[-] BO: Sorting category list tables brings back to root category
In the BO, in Catalog > Categories, if you navigate into a nested
category and try to sort by any of the available parameters, you will
get redirected to the root category, where the sorting will be applied.
You could navigate back to the nested category, but this is not a good
result.