Default tinySetup configuration update
The default tynyMCE configuration for back-office uses only two lines of
buttons, like for CMS pages. But the product admin controller uses the
full version.
I think this complete version should be enabled on CMS pages (and on all
the editor occurences) : the merchant must be able to customize HTML as
he wants.