[*] MO: blockcart animation optimization
The animated element would not be removed upon the animation being finished. This would make the site slow and unresponsive if the animated image was heavy (eg. a transparent png) and added many times without reloading. This fix makes the animation behave as expected also when performing it multiple times.