Crazy name convention in `òverride_folder` and `tpl_folder` from AdminController
Hello,
my Admincontroller Class has the name "RoccExporterController".
At first it is looked after the name convention `<Name>Controller` and the String "Controller" is cutting off from my controller_name and this result in that my controller_name is now "RoccExporter" afterwards the name is truncated from the 5 character until the end of the string but now my controller_name is "xport". I don´t understand this scheme.