Update payment_execution.tpl
With the default bank wire module, if the shop has multiple currencies, it is proposed to the customer only if the merchant accepts payments in all currency. If the merchant accepts payments just in its own currency, the bank wire will not be proposed to the customer that visit the shop with a different currency. That's wrong!

PayPal don't works in this way, the shop can have multiple currencies and it is proposed to everyone but the merchant can receive in its own only.

This mod permits to propose the bank wire in a multiple currencies shop, but add a step that permit to the merchant to receive it in its own.
This is just a "front-end" mod, it needs some work to bring it totally configurable in the BO adding variables, checkboxes, etc. All this is too much for me, I'm not a developer, unfortunately...

In the present mod changes are directly in the code:
- I use "Euros" as default currency (it should be choose in BO among all installed currencies of the shop)
- It works with currencies ID (in this mod: euros=1, dollar=2, pound=3. Everything should follow the installed currencies)

I add some message to help customer in the payment process.