Correctly display "combination unavailable" and hide "add to cart" button in mobile default theme
When an unavailable product combination was selected in the default mobile theme, the JS console reveals an unhandled exception (where the script incorrectly refers to ProductFn.doesntExist as just doesntExist (#442)). This halted the JS execution, preventing the message from being shown, and prevented the "add to cart" button from being correctly hidden.