[-] BO : Fixed position updating issue #PSCSX-2026 PSCSX-1873
A change made on March 24th prevented to update the positions correctly in some tabs (like "Attributes/Values". The $position_group_identifier variable would be always populated with 1 or 0 instead of the $this->position_group_identifier value.

This will make the "id" HTML attribute wrong and methods such as ajaxProcessUpdateAttributesPositions() to fail. Fixed it by keeping the change made on March 24th and using getValue() instead of getIsset().