[-] Installer: console / install theme before modules
We need to install the theme before the module cause they need a theme with a active column to be install. If not, the installation will return false and somes erros will come in console.

If we install the modules before, we will have one errors with the theme.sql process cause table "PREFIX_linksmenutop" doesn't exist at this time.