Update update_order_detail_taxes.php
When I tried to upgrade my prestashop 1.4.6.2 with the "upgrade.php" script, this file "update_order_detail_taxes.php" took about 2 or 3 hours. For each "order_detail" in the database, it create a new "tax" even if its "tax" already exist.
I think the query in the "if($id_tax || $id_tax_alt)" is useless and its result is always false.