[*] CORE : make rewrite/mod_cache to work together
As apache's mod_cache is working *after* mod_rewrite it turns that every "clean url" is mapped to the same front controller.  This makes mod_cache to cache only one URL and giving for every other URL the same output.

This patch append the original "clean path" to the index.php in a fake directory structure making the rewritten URL map one-to-one with the originating URL.

We don't use a GET parameter to not pollute the query string 


If the trick if ok plz apply, otherwise kindly find the time to elaborate an alternative solution, as with or without mod_cache enabled PS performances are drastically different.