Update AdminThemesController.php
I really don't know why the color picker for mails was dropped. It's useful, so I added it again - and I hope the hard coded background-color #f8f8f8 in mails will be changed back to the variable {color} which is still active and usable. Back to the future of 1.5x. ;-)