[-] CORE : actionUpdateQuantity should be called after cache is cleaned
If you are using StockAvailable::getQuantityAvailableByProduct into actionUpdateQuantity hook, results are not "fresh" for the current edited product.
Even if the quantity is real into hook parameters, it may be nice that StockAvailable::getQuantityAvailableByProduct return the fresh result into this hook.
There may be others methods into StockAvailable class that should also work like this