Call to the Dispatcher->getRequiredParams().
May be example will be easier to undersand: if you go to example.com/fr/content/category/1-hoooome, you are redirected to example.com/fr/content/category/1-home?id_cms_category=1, because id_cms_category is not in the excluded_key array. Better is it to fill this excluded_key array dynamically, especially as now there is also the possibility for module to define custom routes through hookModuleRoutes, which may contain any kind of parameters. I'm quite sure of the idea, but quite unsure of the way I've done it...