Update Cart.php
On the back office, in an order made by an employee, when changing the quantity of a product line, the last modified product go to the bottom of the list. This is problematic.

The problem come from the order by the column date_add in database recently updated in the function getProducts of the same class (between 1.6.0.8->1.6.0.9, the order was from the product id and now it is from the creation date). 

Firstly it make a the previous problem that I said, and secondly its pretty weird to update the date of creation when updating a product cart