Change bulk actions drop-DOWN to drop-UP
Are you tired of dropdowns, that appears at very bottom of page, so when you click on it, it opens away from your screen and you have to scroll down, because you don't see the options? Well suffer no more, because we can change this behavior by single class assignment. Thanks to the bootstrap, we can add class:dropup on parent element and the dropdowns now appear in visible part of screen. This will in fact improves your efficiency in back-office as well as quality of your overall life and it's free! ;)