[-] FO : Display subject description in shop contact form on page load
It was possible to provide the argument "id_contact" for the shop contact form page to pre-select a suject on the form. The subject heading was correctly pre-selected but its description was not displayed until it was re-selected. The description of the form subject is now displayed on page load.