Language deletion not complete
It seems there was an error in the regexp used to identify language related tables (no table was processed).
In addition, database deletion errors are now ignored in order not to fail when tables from additional modules don't respect the id_lang column convention (e.g. Affinity Items v2.0.0). This is to be discussed.