Update StockAvailable.php
I'm currently developing a module which send alert by mail when quantity is changing for products. 
I have noticed that the hook "actionUpdateQuantity" is not properly set if my database contains many shops which sell a same product. 
It's impossible to make a difference between two execution of "actionUpdateQuantity" without the id_shop attribute for a same product.