Fix unity validation
The unity string is of type VARCHAR(255) in the database, but the client-side validation restricts it to 10 characters.
The CSV-Import, however, does not. When the user tries to edit an imported product with a unity string of more than 10 characters, the unity text field is initialized with the too long / invalid string and the product can not be saved.
This commit fixes this by allowing a longer unity string.