[-] FO : Fixed bug showing errors in manufacturer list
When non of the products doesn't have association with any of the manufacturers, then the FO manufacturer list throws error whenever it checks if the number of products associated with the current brand is greater than 0(zero). I've added additional check in front every of this checks, if the number of products exists at first.