Added a poor's man javascript hook system
This is a proof of concept and demostrate how deeply nested components
may be given some modularity and extensibility easily anyway.
In this case it allows any module to insert it's own code before
client-cart-related data to a new cart addition are submitted to the
Cart handler.

Eg: a module adding a new ```<input>``` may implement said hook in order
to add ```$(newfield).val()``` to the list of parameters passed to an
(hypothetically overriden) Cart controller.

Eg: in mymodule.php:
```php
public function hookActionFrontControllerSetMedia() {
   $this->context->controller->addJS($this->_path . 'js/cart-summary-mymodule.js');
   $JsDefs = Media::getJsDef();
   // don't override if already other function already hooked it
   $curCartHooks = isset($JsDefs['hookListingShoppingCartPreAdd']) ? $JsDefs['hookListingShoppingCartPreAdd'] : array();
   $curCartHooks[] = 'hookShoppingCartPreAdd_updateXXXX';
   Media::addJsDef(array('hookListingShoppingCartPreAdd' => $curCartHooks));
}
```

In js/cart-summary-mymodule.js:
```javascript
// preprocess the datastring sent to the cart controller
function hookShoppingCartPreAdd_updateXXXX(dataArray) {
	if(the-condition-we-want) {
		dataArray.ourvariable = parseFloat($('#some-new-stuff-to-send-to-the-cart-controller').val());
	}
	return dataArray;
}
```

Then hookShoppingCartPreAdd_updateXXXX() will be run with "args" as its first and unique parameter.
It's returns value will be given to posterior hooks before being sent to the Cart controller as a query-string.