Avoid using stale Entity information indirectly from the database cache.
The EntityMapper first tries to fetch Entity (partial or full) from a cache_id.
(eg: objectmodel_Product_223_1_0)

But if that fails it will try to fetch information using Db::getRow() which, by
default use it's own cache. In such situation stale results are likely to be
provided. Since database results are intended to be stored, let's *really* fetch
them from the MySQL database.

Real-case: Using memcached cache it was found that the back-office was providing
old result after a form submission (eg: priceTE input value would be be restored
to its initial value although actually updated in the database).
This was tracked down to this scenario where memcached had kept old mysql result
for which the invalidation was almost impossible.