[-] FO: Bugfixes in store locator
This PR prevents problem with searching for store with specific postcode, for example 33181 can be in Germany, Ukraine and USA, this change prevents all the problems with this kind of situation.

How to reproduce this bug? Search for eg. 33181 in store locator, results array from Geocoder has few elements but currently in script we had results[0].