[-] FO : Re-instate default id
Fixes broken blocklayered ajax product-list refresh and maybe other modules that don't pass the $id parameter to the tpl. As a reminder, the long standing id was pushed into the class list, then it was added again as a parameter but conditionally and without a default. There are other ways to fix this but not as easily while maintaining compatibility (refactoring blocklayered.js like on demo.ps.com doesn't seem right nor necessary). It is not blocklayered's fault that this happened.

PS: Do I need to push this anywhere else as well?