[*] FO : order email product cust.: adjust to FO display
Customizations don't change the price of a product and are displayed in
the FO cart without the extra product name and price. Especially
duplicating the price in the confirmation email confuses the customer.

This change makes the displayed information in the FO cart and the
product list in the order confirmation email the same.

(There's no commit category for email templates, I hope FO is ok.)