[*] CORE : FO Search position considering stocks
Even if products are listed by position (calculated by word matching), we must consider that this is not an absolute value.

For example : I search for a brand name with ajax search, only 10 (by default) products are shown, which is not much, customer should see available products. If 10 first products are out of stock and not the one after, this one should be positionned first.

We should also consider number of element shown by pages for ajax search instead of 10, search results are defined through templates, taking care of. 

I also used PS_PRODUCTS_PER_PAGE for classic searches.