[*] CORE : Make missing "blocklayered" product fields available in template
These fields are important as they are used as part of the product when blocklayered is active. They are associated with features which are already used in the template, but the additional fields can't be used for enhancing product.tpl as they are not exported. Elsewhere in the core product class and controller, the existence of blocklayered features are already checked and integrated. Modifications tested for months in production environments and work.