[-] Bugfix HTMLTemplateInvoice.php
This is nuts, guys. It makes absolutely no sense to store the invoice and delivery address sort of 'hard-coded' to ps_order_invoice. It happens nearly every day that a merchant has to change the invoice or delivery address of a placed order and regenerate the bill. 
Up to 1.6.0.14 this worked like a charm, you changed the customer address from the BO order page and regenerated the invoice. From 1.6.1 these changes affect the customer's address, but no longer the current invoice. 
This is no practical solution but screams for a redesign. ;)