[-] BO
if you want to add an order via backoffice and you have chosen an existing customer, his carts and orders are loaded. But if the address of the cart is not existing anymore, there is no chance to edit it. This code change makes sure, that an existing backup address is loaded properly.