[*] FO Shopping Cart Change
When on the cart page there is a Products with a total in the right block that in misleading and unnecessary as it actually show the Grand Total of the products with shipping and discounts if applicable.
I find the line unnecessary as this total is already shown on the Shopping Cart so I changed to a title Shopping Summary I also changed this is the steps pages.
I also added some CSS to give a bottom margin, line 6508 theme.css

I also noticed that once on the steps pages the Tax was shown in bold as it seems this was missed in the CSS file so I corrected at line 5225 theme.css.
I also noticed that once on the steps pages that if you have a discount it does not have a top margin this to seems to have bin missed in the CSS so I corrected line 5211 theme.css.