[*] BO Shop Link Change
After fresh install of 1.7 alpha 3 I noticed that the Shop link would open in the current window instead on a new tab this change corrects that.
I also noticed that after some time or possibly some BO change that this link would change to the header.tpl and that pull request was made by another member.
This file may have something to do with Multi-Shop but not sure as I do not use.