[*] FO Cart Summary Change
When on the cart summary page there is a line showing the Products with a total at the top of the right block that in misleading and unnecessary as it actually show the Grand Total of the products with shipping and discounts.
I find the line unnecessary as this total is already shown on the Shopping Cart so I changed to the title Shopping Summary I also changed this is the steps pages.
I also added some CSS to give a bottom margin.

In the steps pages only the Tax was shown in bold as it seems this was missed in the CSS file so this corrects that, I also removed the font-weight to keep the text more in line with the cart summary page.
In the the steps pages if you have a discount the discount line is missing a top margin this will correct that.
I also added an HR to separate the summery from the total.