[*] FO Product Image
When switching to a 2 column layout on the product page the right column description area appears to overlap the image this is due to the product cover and hover backgrounds not scaling to the window width.
This change will scale both to fit the scaled image and works in all layouts.