[*] FO Checkout Steps
When checking out on summary block tax line has the tax in bold not matching the other totals labels this seems to have been missed from the CSS and I updated the font weight to better match the cart summary page which is much bolder.
If you add a voucher code the Discount line has to top margin this seems to have been missed from the CSS.
There is no divide between the summary lines and the Total I added an HR here, this is the best way to place a divider and matches all other divides found in the other cart summary tpl files.
I hope none of this is seen as changing style as that is not the intention but to enhance the appearance.