[-] Install
1.7.0.0 Alpha 4 fails during install with this message
An SQL error occurred for entity theme_meta: Table 'xxxxxx_shop17test.ps_theme_meta' doesn't exist

I also notice that in the XML folder there is a theme.xml that points to default-bootstrap I had also added this table as well but removed because the theme.xml points to default-bootstrap which appears to be 1.6.

The install works with this fix.