B2B with internationalization
SIRET and APE is not a feature with international use, then must be and abstract method to allow all countries to use it only by overriding this class. Then I suggest you to delete this field in Customer class and implements as a module and an apart class in future versions.