Update Tools.php (function setCurrency)
New code allows to change the default currency display when person is firstly visiting shop. The old code assigned the Default Displayed Currency for the first visiting equal to those Currency, which is the one prices of products are indicated (also in DB). The new code allow to change Currency shown during the first visit. This would be usefull if the prices are in dollars, for intance, and the shop is made for Russian audience: it is more convenient to show price in local currency at once. New Config value 'PS_DISPLAY_DEFAULT_CURRENCY'  will allow to make module controlling the Default Displayed Currency without overriding Tools.php