Update Media.php
Check if 'HTTP_USER_AGENT' is defined before trying to use it.  If a user agent is not provided in the request, the original code leads to a warning in the php error log.  
stderr: PHP Notice:  Undefined index: HTTP_USER_AGENT in /var/www/vhosts/host/httpdocs/shop/classes/Media.php on line 604