CO: Faster loading pages combination product
This change greatly speeds the load tab "combinations" on page
"product". This is particularly evident if any group attributes a large
number of values. In the old cycle, at each iteration in the array is
placed a new value attribute and the sorting of the array. If values,
e.g., 1000, the array is sorted 1000 times! This is terrible!
After 1 times is enough.