Update Dispatcher.php
Url with a filter like monsite.fr/3-femmes/Couleur-Bleu doesn't show filtered products. So on the dispatch, i see that the param configureted was "selected_filter", but the param use in the url is "q". When i made the change, the filtering work properly. I don't know if is the right way to debug this feature.