FO: Fix ecotax rounded twice
If taxes are turned on on ecotax, ecotax used to be rounded twice. Once after fetching the DB value and the second time after multiplying it with the tax rate. Now it's only rounded after taxes.
Example: If you want an ecotax of 10 (with 21% VAT), the ecotax in database is 8.264463 (because 8.264463 * 1.21 = 10). But without this fix, PrestaShop would round the ecotax DB value to 8.26, then multiply it 8.26 * 1.21 = 9.99 (wrong, one cent too little)