FO: Prevent empty cart of death
In specifics conditions , customer can't set/update his cart :
* In FO, to each product added, cart display stay empty.
* In BO, each action is logged as a new cart with an the product added.

On add action, cookie status is not updated and id_cart is not related to add action. It's explain these behaviors.

To prevent this we must save as soon as possible the id cart in cookie. Cookie write action in DisplayAjax can be act too late or never.
Then when a new cart is set, we save this id right now.
Looks better also to save cookie before any dieAjax to insure cookie consistency.