BO: Fix coverage calculations in Stock Cover Report
When caclulating the coverage qty per day this value is prematuely rounded using the shops currency rounding (ps_round) resulting in infinite qty coverage result.

If the shop is using the default rounding any coverage calculation where the result of the calculation is less than 0.5 will result a very inaccurate coverage calculation.

For example: Lets assume you have a product that sold 14 times ($quantity_out) over a coverage period of one month (30 days) ($coverage).  The result of this calc is (14/30) = 0.4667 which is then rounded down to zero.

This ends with most coverages reporting as infinite when in actual fact there is significant impact on stock coverage.