UNIX_TIMESTAMP converts to UTC time
UNIX_TIMESTAMP convert MySQL dates to UTC (https://dev.mysql.com/doc/refman/5.5/en/date-and-time-functions.html#function_unix-timestamp). You should not use that while comparing it to a system date, if the system timezone is different from UTC. I had problems with cache because of that, and the cache was deleted at every refresh. Took me a while to find the problem.