default AJAX Relevance search sort order is wrong
search for "summer" in demo store. All products with the "summer" in the product name should appear at the first lines in the AJAX search results, but I see them at the 3rd and 4th positions (results shown immediately but in the ASC relevance order instead of DESC). This error is continuation of the pull request #8399. The issue was not fixed in full. as an addition to the fix in pull request #8399 I suggest the following change in the controllers/front/listing/SearchController.php: ->setSortOrder(new SortOrder('product', Tools::getProductsOrder('by', 4), Tools::getProductsOrder('way', 1))) instead of ->setSortOrder(new SortOrder('product', Tools::getProductsOrder('by'), Tools::getProductsOrder('way')))