Wrong permissions for image folders
Image folder is being created with wrong permissions, and chmod works only on the last directory. chmod does not recursively change permissions, so mkdir makes dirs with wrong permissions and chmod only changes the last one. Ive seen this solution within PrestaShop files elsewhere, but it is only a suggestion.