Modify CustomerAddressFormatter.php
With PS 1.7.2.4, if DNI is enabled in Customers -> Address, and not enabled in Countries -> countries where DNI is not mandatory, a 500 error occurs for these countries after submitting of the address. It's necessary to unable DNI in Customers -> Address and enable it in Countries only for these where the DNI is mandatory (like Italy, ...). For these countries, "Optional" cannot be display near this field. So this code get this field Required for these countries.