Use Symfony's request to build request URI
Leave it optional to inject a Request object (to not break existing code) to the dispatcher. If not injected create one and save it in a property. This way the Dispatcher has access to the Symfony Request object and can build the request URI from it (and is able to use it somewhere else too or pass it around)