CO: fixes validation of a hook name when it contains forward slashes
Rationale:
When saving an Object Model, the actionObject[className]BeforeSave hook is triggered.
For Object Model classes defined within a namespace, the name is the fully qualified namespace
including forwards slashes and breaking the hook name validator regex.

The proposed fix allows for forward slashes to be part of hook names.