Social title field
If genders gets deleted dont show empty form field. This will allow to hide genders without editing code. Option for social title like "ask for birth date" would be great, but this is still required until there is a capability of social title deletion. I see this as a bug. Should be cool to have an option for this in the future. Because every time I need to remove this.