Property Address->id_country is empty
When trying to delete old address called from CustomerAddressPersister.php:73 "return $address->save() && $old_address->delete();" Fatal error is thrown about id_country being empty. This is because old address does not exists and at the same time prestashop tries to validate it and update it and then throws error. At this case if ID is empty means records is not present on database and shouldnt be updated nor deleted. And there is a another bug that isUsed() returns true on non persisted address record. But in any way if ID is empty no further actions should be performed.