Fix for browser back button on product page
If you go to product page from product list, then if you switch attributes 3 times, then you need to 3 times click on back button to go back to product list page which is not very user friendly. This Pr fix that and only one click on back button is needed.