Add missing appendArray in OrderReturnLazyArray (BOOM-6039)
The FO page presenting the OrderReturn did not work anymore due to
modifications on the OrderReturnPresenter::present method which no
loner returns an array but a OrderReturnLazyArray object which is
perfect to access the generated fields like detail_url through the
getDetailsUrl method. But it removed to access to original array
data like state_name and such.

Rather than implementing all the getter methods we simply use the
AbstractLazyArray::appendArray method to open access to all the
missing array fields.