Propose improvement
First, get baselink only once, instead of each time -> could have a very slight speed improvement.
Second, the if(strpos($quickAccess[$index]['link'], 'http') !== 0 or strpos($quickAccess[$index]['link'], $baselink) === 0) is used to allow other links, like the https://google.com, etc... It only adds the baselink and token to the link that do not start with http (or https of course) or to those that start with http, but it's the same as the current used link. If it's another link for the same shop, it will ask for login any way.
Third, moved the check to add the admin token to inside the else.  This way it doesn't add to store url.
Fourth, moved the clean url, to be inside the else to generate the shop front office url when link is '../'.