Fix a bug when you try to update a carrier range.
When you try to update a range and the delimiters existed previously, on a carrier with the same id_reference, the range is not updated and the carrier keeps the current ranges.

Now, instead of only checking if a range exists, we get the id of the range with the provided delimiters. The id of the range is used to load the range object and assign to it the id of the updated carrier.