Delivery price cannot be defined by shop
The context was forced to "All shops" with this : $this->multishop_context = Shop::CONTEXT_ALL;
I've tried to remove this and it works fine : you can define a delivery price for a specific shop. I've tested on front side and the correct price is applied.
This fix need to be more intensively tested.