[BO] Verify stock if needed before updating to delivered
There is a stock issue occurring if you follow these steps :

- On a Prestashop with Advanced Stock Management and Back-order allowed
- Create two warehouses A and B
- Create a product P that can be stored in both warehouses
- Create an order with 2 P
- Update the order status to payment accepted ( the warehouse A is the default one, so it's real quantity goes to -2 )
- When your supplier deliver you the 2 P, you make a mistake and increase the quantity in the warehouse B
- Update the order status to delivered

The warehouse A usable and real quantity are now at 0, and the warehouse B usable and real quantity are at 2, resulting in an overall quantity of 2, which is wrong.

One can think that the physical quantity of the warehouse A should be at -2, resulting in an overall quantity of 0. But that is just wrong, we can't have negative stock value in the physical world. 0 is the correct value here.

The problem is that when removeProduct() is called, there is no check for success, it is just assumed that it was done correctly.

As there is not enough quantity in warehouse A to deliver the products, it should not let us update the status to delivered.

This commit fix that by checking if the quantity can be removed, if necessary, before allowing the status update to delivered.

In this case an error will be displayed and a transfer of 2 P from warehouse B to A will be necessary before being able to deliver the product.